Avoid importing "e2e CLI" package into the "e2e test" package
